*: move Cluster interface to api
